make document a component
