+++ app/components/document/template.html
@@ -1,4 +1,4 @@
-<div ng-controller="DocumentCtrl">
+<div>
   <div class="ph-header ph-header-large ph-lg-margin-bottom">
     <div class="container-fluid">
       <div class="row">
